Adding new scrapy project : ScrapyAndroid
Works without being blocked by google servers. Using a counter to limit the number of pages to scrape. The extracted data is stored in the file 'android.csv'

The extracted fields are defined and processed in items.py.
In setting.py, we define all our parameters for scraping, like the format export of the csv file, or the language to use while scraping.